requirements.txt
asgiref==3.10.0
boto3==1.40.64
botocore==1.40.64
Django==5.2.7
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
jmespath==1.0.1
pillow==12.0.0
postgres==4.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
psycopg2-pool==1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
